completed level1/AddDigit 2020-09-17 13:15

+++ ProgPRJ/src/level1/SumBtw2Ints.java
@@ -13,7 +13,7 @@ public class SumBtw2Ints {
 		if(a > b) {
 			swNsum=a;
 			a=b;
-			b=(long)swNsum;
+			b=(int)swNsum;
 			swNsum=0;
 		}
 		for(int i=a; i <= b; i++)
